requirements.txt
fastapi
uvicorn[standard]
psycopg[binary,pool]
Pillow
